+++ ignite/cmd/scaffold_chain.go
@@ -80,6 +80,7 @@ about Cosmos SDK on https://docs.cosmos.network
 	c.Flags().StringP(flagPath, "p", "", "create a project in a specific path")
 	c.Flags().Bool(flagNoDefaultModule, false, "create a project without a default module")
 	c.Flags().StringSlice(flagParams, []string{}, "add default module parameters")
+	c.Flags().StringSlice(flagModuleConfigs, []string{}, "add module configs")
 	c.Flags().Bool(flagSkipGit, false, "skip Git repository initialization")
 	c.Flags().Bool(flagMinimal, false, "create a minimal blockchain (with the minimum required Cosmos SDK modules)")
 
@@ -97,14 +98,16 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		noDefaultModule, _ = cmd.Flags().GetBool(flagNoDefaultModule)
 		skipGit, _         = cmd.Flags().GetBool(flagSkipGit)
 		minimal, _         = cmd.Flags().GetBool(flagMinimal)
+		params, _          = cmd.Flags().GetStringSlice(flagParams)
+		moduleConfigs, _   = cmd.Flags().GetStringSlice(flagModuleConfigs)
 	)
 
-	params, err := cmd.Flags().GetStringSlice(flagParams)
-	if err != nil {
-		return err
-	}
-	if noDefaultModule && len(params) > 0 {
-		return errors.New("params flag is only supported if the default module is enabled")
+	if noDefaultModule {
+		if len(params) > 0 {
+			return errors.New("params flag is only supported if the default module is enabled")
+		} else if len(moduleConfigs) > 0 {
+			return errors.New("module configs flag is only supported if the default module is enabled")
+		}
 	}
 
 	cacheStorage, err := newCache(cmd)
@@ -123,6 +126,7 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		skipGit,
 		minimal,
 		params,
+		moduleConfigs,
 	)
 	if err != nil {
 		return err