+++ ignite/services/chain/lint_test.go
@@ -1,0 +1,6 @@
+package chain_test
+
+import "testing"
+
+func TestChainLint(t *testing.T) {
+}